Rename the mogonet trainer
